+++ src/main/java/com/example/kkbackend/dtos/MemberDto.java
@@ -3,5 +3,5 @@ package com.example.kkbackend.dtos;
 import lombok.Builder;
 
 @Builder
-public record MemberDto(Integer telegramId, String username, String firstName, boolean freshBlood) {
+public record MemberDto(Double telegramId, String username, String firstName, boolean freshBlood) {
 }
