+++ src/main/java/com/example/kkbackend/dtos/RateDto.java
@@ -3,5 +3,5 @@ package com.example.kkbackend.dtos;
 import lombok.Builder;
 
 @Builder
-public record RateDto(String id, Integer telegramId, String username, String firstName, String movieId, int rating, boolean liked, boolean discussable) {
+public record RateDto(String id, Double telegramId, String username, String firstName, String movieId, int rating, boolean liked, boolean discussable) {
 }
